+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * Multiplayer:
   * Fixed effect desync issues, for real this time, probably.
 
+* All Interactables Are Cloaked:
+  * Default duration: 60s -> 90s
+
 * Increase/Decrease Effect Duration:
   * Duration multiplier now applies to the remaining time rather than the total duration. Stops decreased effect duration ending other effects immediately.